Add mock for AsyncStorage in Jest tests and update Home and Search tests
