+++ __mocks__/@react-native-async-storage/async-storage.js
@@ -1,0 +1,1 @@
+module.exports = require('@react-native-async-storage/async-storage/jest/async-storage-mock');
